move database functions to new module
